Refactor iqk: Factor out float GEMM (AVX2/AVX512)
